Move Ecc tests to their own file
