+++ src/tests/mod.rs
@@ -213,93 +213,6 @@ mod attrs;
 
 mod objects;
 
-#[test]
-fn test_create_ec_objects() {
-    let mut testtokn =
-        TestToken::initialized("test_create_ec_objects.sql", None);
-    let session = testtokn.get_session(true);
-
-    /* login */
-    testtokn.login();
-
-    let mut handle: CK_ULONG = CK_INVALID_HANDLE;
-
-    let mut class = CKO_PUBLIC_KEY;
-    let mut ktype = CKK_EC;
-    let mut verify: CK_BBOOL = CK_TRUE;
-    let label = "EC Public Signature Key";
-    let point_hex = "041b803bf0586decf25616e879b0399aa3daab60916fc76c9b6c687fc1454cba90d5f15aeb36e7070cffb4966499b71b389453c0075203fa047d4f3e44343edc84fb793bf1b8ca94dd3f293afbe68e3be93f1245be9fb71be3c50f1263bc12d516";
-    let params_hex = "06052b81040022";
-    let point = hex::decode(point_hex).expect("Failed to decode hex point");
-    let params = hex::decode(params_hex).expect("Failed to decode hex params");
-    let mut template = vec![
-        make_attribute!(CKA_CLASS, &mut class as *mut _, CK_ULONG_SIZE),
-        make_attribute!(CKA_KEY_TYPE, &mut ktype as *mut _, CK_ULONG_SIZE),
-        make_attribute!(CKA_VERIFY, &mut verify as *mut _, CK_BBOOL_SIZE),
-        make_attribute!(
-            CKA_LABEL,
-            label.as_ptr() as *mut std::ffi::c_void,
-            label.len()
-        ),
-        make_attribute!(
-            CKA_EC_POINT,
-            point.as_ptr() as *mut std::ffi::c_void,
-            point.len()
-        ),
-        make_attribute!(
-            CKA_EC_PARAMS,
-            params.as_ptr() as *mut std::ffi::c_void,
-            params.len()
-        ),
-    ];
-
-    let ret = fn_create_object(
-        session,
-        template.as_mut_ptr(),
-        template.len() as CK_ULONG,
-        &mut handle,
-    );
-    assert_eq!(ret, CKR_OK);
-
-    /* Private EC key */
-    class = CKO_PRIVATE_KEY;
-    let mut ktype = CKK_EC;
-    let mut sign: CK_BBOOL = CK_TRUE;
-    let label = "EC Private Signature Key";
-    let value_hex = "4a77d1245d2c4751ff178040cc9e527b4d6cbb067b8fb01265b854fa581fd62dadc706025cbf515d80fd226f8f552f34";
-    let value = hex::decode(value_hex).expect("Failed to decode value");
-    template = vec![
-        make_attribute!(CKA_CLASS, &mut class as *mut _, CK_ULONG_SIZE),
-        make_attribute!(CKA_KEY_TYPE, &mut ktype as *mut _, CK_ULONG_SIZE),
-        make_attribute!(CKA_SIGN, &mut sign as *mut _, CK_BBOOL_SIZE),
-        make_attribute!(
-            CKA_LABEL,
-            label.as_ptr() as *mut std::ffi::c_void,
-            label.len()
-        ),
-        make_attribute!(
-            CKA_VALUE,
-            value.as_ptr() as *mut std::ffi::c_void,
-            value.len()
-        ),
-        make_attribute!(
-            CKA_EC_PARAMS,
-            params.as_ptr() as *mut std::ffi::c_void,
-            params.len()
-        ),
-    ];
-
-    let ret = fn_create_object(
-        session,
-        template.as_mut_ptr(),
-        template.len() as CK_ULONG,
-        &mut handle,
-    );
-    assert_eq!(ret, CKR_OK);
-
-    testtokn.finalize();
-}
-
 mod init;
 
 mod mechs;
@@ -310,181 +223,7 @@ mod rsa;
 
 mod session;
 
-#[test]
-fn test_ecc_operations() {
-    let mut testtokn = TestToken::initialized(
-        "test_ecc_operations.sql",
-        Some("testdata/test_ecc_operations.json"),
-    );
-    let session = testtokn.get_session(true);
-
-    /* login */
-    testtokn.login();
-
-    /* private key */
-    let mut handle: CK_ULONG = CK_INVALID_HANDLE;
-    let mut template = vec![make_attribute!(
-        CKA_UNIQUE_ID,
-        CString::new("5").unwrap().into_raw(),
-        1
-    )];
-    let mut ret = fn_find_objects_init(session, template.as_mut_ptr(), 1);
-    assert_eq!(ret, CKR_OK);
-    let mut count: CK_ULONG = 0;
-    ret = fn_find_objects(session, &mut handle, 1, &mut count);
-    assert_eq!(ret, CKR_OK);
-    assert_eq!(count, 1);
-    assert_ne!(handle, CK_INVALID_HANDLE);
-    ret = fn_find_objects_final(session);
-    assert_eq!(ret, CKR_OK);
-
-    /* sign init */
-    let mut mechanism: CK_MECHANISM = CK_MECHANISM {
-        mechanism: CKM_ECDSA,
-        pParameter: std::ptr::null_mut(),
-        ulParameterLen: 0,
-    };
-    ret = fn_sign_init(session, &mut mechanism, handle);
-    assert_eq!(ret, CKR_OK);
-
-    /* a second invocation should return an error */
-    ret = fn_sign_init(session, &mut mechanism, handle);
-    assert_eq!(ret, CKR_OPERATION_ACTIVE);
-
-    let data = "plaintext";
-    let sign: [u8; 64] = [0; 64];
-    let mut sign_len: CK_ULONG = 64;
-    ret = fn_sign(
-        session,
-        CString::new(data).unwrap().into_raw() as *mut u8,
-        data.len() as CK_ULONG,
-        sign.as_ptr() as *mut _,
-        &mut sign_len,
-    );
-    assert_eq!(ret, CKR_OK);
-    assert_eq!(sign_len, 64);
-
-    /* a second invocation should return an error */
-    ret = fn_sign(
-        session,
-        CString::new(data).unwrap().into_raw() as *mut u8,
-        data.len() as CK_ULONG,
-        sign.as_ptr() as *mut _,
-        &mut sign_len,
-    );
-    assert_eq!(ret, CKR_OPERATION_NOT_INITIALIZED);
-
-    /* test that signature verification works */
-    template = vec![make_attribute!(
-        CKA_UNIQUE_ID,
-        CString::new("4").unwrap().into_raw(),
-        1
-    )];
-    ret = fn_find_objects_init(session, template.as_mut_ptr(), 1);
-    assert_eq!(ret, CKR_OK);
-    let mut count: CK_ULONG = 0;
-    ret = fn_find_objects(session, &mut handle, 1, &mut count);
-    assert_eq!(ret, CKR_OK);
-    assert_eq!(count, 1);
-    assert_ne!(handle, CK_INVALID_HANDLE);
-    ret = fn_find_objects_final(session);
-    assert_eq!(ret, CKR_OK);
-
-    ret = fn_verify_init(session, &mut mechanism, handle);
-    assert_eq!(ret, CKR_OK);
-
-    ret = fn_verify(
-        session,
-        data.as_ptr() as *mut u8,
-        data.len() as CK_ULONG,
-        sign.as_ptr() as *mut u8,
-        sign_len,
-    );
-    assert_eq!(ret, CKR_OK);
-
-    /* P-521 private key */
-    let mut handle: CK_ULONG = CK_INVALID_HANDLE;
-    let mut template = vec![make_attribute!(
-        CKA_UNIQUE_ID,
-        CString::new("7").unwrap().into_raw(),
-        1
-    )];
-    ret = fn_find_objects_init(session, template.as_mut_ptr(), 1);
-    assert_eq!(ret, CKR_OK);
-    let mut count: CK_ULONG = 0;
-    ret = fn_find_objects(session, &mut handle, 1, &mut count);
-    assert_eq!(ret, CKR_OK);
-    assert_eq!(count, 1);
-    assert_ne!(handle, CK_INVALID_HANDLE);
-    ret = fn_find_objects_final(session);
-    assert_eq!(ret, CKR_OK);
-
-    /* sign init */
-    let mut mechanism: CK_MECHANISM = CK_MECHANISM {
-        mechanism: CKM_ECDSA,
-        pParameter: std::ptr::null_mut(),
-        ulParameterLen: 0,
-    };
-    ret = fn_sign_init(session, &mut mechanism, handle);
-    assert_eq!(ret, CKR_OK);
-
-    /* a second invocation should return an error */
-    ret = fn_sign_init(session, &mut mechanism, handle);
-    assert_eq!(ret, CKR_OPERATION_ACTIVE);
-
-    let data = "plaintext";
-    let sign: [u8; 132] = [0; 132];
-    let mut sign_len: CK_ULONG = 132;
-    ret = fn_sign(
-        session,
-        CString::new(data).unwrap().into_raw() as *mut u8,
-        data.len() as CK_ULONG,
-        sign.as_ptr() as *mut _,
-        &mut sign_len,
-    );
-    assert_eq!(ret, CKR_OK);
-    assert_eq!(sign_len, 132);
-
-    /* a second invocation should return an error */
-    ret = fn_sign(
-        session,
-        CString::new(data).unwrap().into_raw() as *mut u8,
-        data.len() as CK_ULONG,
-        sign.as_ptr() as *mut _,
-        &mut sign_len,
-    );
-    assert_eq!(ret, CKR_OPERATION_NOT_INITIALIZED);
-
-    /* test that signature verification works */
-    template = vec![make_attribute!(
-        CKA_UNIQUE_ID,
-        CString::new("6").unwrap().into_raw(),
-        1
-    )];
-    ret = fn_find_objects_init(session, template.as_mut_ptr(), 1);
-    assert_eq!(ret, CKR_OK);
-    let mut count: CK_ULONG = 0;
-    ret = fn_find_objects(session, &mut handle, 1, &mut count);
-    assert_eq!(ret, CKR_OK);
-    assert_eq!(count, 1);
-    assert_ne!(handle, CK_INVALID_HANDLE);
-    ret = fn_find_objects_final(session);
-    assert_eq!(ret, CKR_OK);
-
-    ret = fn_verify_init(session, &mut mechanism, handle);
-    assert_eq!(ret, CKR_OK);
-
-    ret = fn_verify(
-        session,
-        data.as_ptr() as *mut u8,
-        data.len() as CK_ULONG,
-        sign.as_ptr() as *mut u8,
-        sign_len,
-    );
-    assert_eq!(ret, CKR_OK);
-
-    testtokn.finalize();
-}
+mod ecc;
 
 #[test]
 fn test_hashes_digest() {
